Fix the random sampling when plotting .data. files
This change fixes a bug in which the random sample used to plot data
files is then used to calculate the frontal position, rather than using
all particles.

Also, you can now pass a sample size that is a floating-point number
less than 1, which will be interpreted as the fraction of particles that
should be plotted.

+++ mdpmflc/controller/results/plots.py
@@ -50,7 +50,7 @@ def showdataplot_fig(sername, simname, ind, form="png"):
         os.makedirs(os.path.dirname(dataplot_fn), exist_ok=True)
 
         samplesize = flask.request.values.get("samplesize")
-        samplesize = int(samplesize) if samplesize else 20000
+        samplesize = float(samplesize) if samplesize else 20000
 
         width = flask.request.values.get("width")
         width = float(width) if width else 7
